tests: ceedling: reorganize
tests: ceedling: fix per-suite lib build staleness and races

Use a separate CARGO_TARGET_DIR per suite so parallel make cannot
overwrite target/debug/libsmart_keymap.a mid-build. Track smart_keymap
Rust/ncl inputs as prerequisites so copied libs rebuild when sources
change, not only when keymap.ncl changes.

tests: ceedling: fix per-suite link flags for CI

Ceedling 1.0.0 matches :flags :test :link matchers against the test
filename, not the suite subdirectory path. Switch to per-test matchers
and link the copied archive paths directly.

tests: ceedling: link per-suite libs via mixins on Linux

:flags :test :link matchers prepend libraries before objects, which
fails on Linux with static archives. Run ceedling once per suite with
a mixin that sets :libraries (correct link order), using test:path[].

tests: ceedling: shorten suite guard #error messages

Drop path and fixture filename from #error text so clang-format
keeps them on one line.
