+++ tests/ceedling/.gitignore
@@ -1,0 +1,3 @@
+libs/
+generated/
+cargo-target/
