+++ tests/ceedling/test/tap_hold/test_keydef_taphold_event_driven.c
@@ -1,3 +1,7 @@
+#include "tap_hold_test_ceedling_fixture.h"
+
+#ifdef SUITE_TAP_HOLD
+
 #include "unity.h"
 
 #include "hid_keycodes.h"
@@ -7,43 +11,6 @@ void setUp(void) {}
 
 void tearDown(void) {}
 
-// KEYMAP: [C & TH.LCtrl, D & TH.LSft, A, B]
-
-void test_event_driven_key_press(void) {
-  uint8_t expected_report[8] = {0, 0, KC_A, 0, 0, 0, 0, 0};
-  KeymapHidReport report = {};
-  KeymapHidReport *actual_report = &report;
-
-  keymap_init();
-
-  keymap_register_input_after_ms(
-      0, (struct KeymapInputEvent){.event_type = KeymapEventPress, .value = 2},
-      actual_report); // Third key in the keymap is A
-
-  TEST_ASSERT_EQUAL_UINT8_ARRAY(expected_report, actual_report->keyboard, 8);
-}
-
-void test_event_driven_key_tap(void) {
-  uint8_t expected_report[8] = {0, 0, 0, 0, 0, 0, 0, 0};
-  KeymapHidReport report = {};
-  KeymapHidReport *actual_report = &report;
-
-  // assemble: init keymap
-  keymap_init();
-
-  // act: press then release A via event-driven API
-  keymap_register_input_after_ms(
-      0, (struct KeymapInputEvent){.event_type = KeymapEventPress, .value = 2},
-      actual_report); // Third key in the keymap is A
-  keymap_register_input_after_ms(
-      0,
-      (struct KeymapInputEvent){.event_type = KeymapEventRelease, .value = 2},
-      actual_report); // Third key in the keymap is A
-
-  // assert: empty report
-  TEST_ASSERT_EQUAL_UINT8_ARRAY(expected_report, actual_report->keyboard, 8);
-}
-
 void test_event_driven_tap_hold_key_tap(void) {
   KeymapHidReport report = {};
   KeymapHidReport *actual_report = &report;
@@ -51,15 +18,12 @@ void test_event_driven_tap_hold_key_tap(void) {
   // assemble: init keymap
   keymap_init();
 
-  // Tap the C & TH.LCtrl
-  uint16_t keymap_index = 0;
-
   // act: press tap-hold key at t=0
   {
     uint32_t actual_next_ev_ms = keymap_register_input_after_ms(
         0,
         (struct KeymapInputEvent){.event_type = KeymapEventPress,
-                                  .value = keymap_index},
+                                  .value = KM_TAP_HOLD_KEY},
         actual_report);
 
     // assert: hold timeout scheduled, no key output yet
@@ -73,7 +37,7 @@ void test_event_driven_tap_hold_key_tap(void) {
     uint32_t actual_next_ev_ms = keymap_register_input_after_ms(
         150,
         (struct KeymapInputEvent){.event_type = KeymapEventRelease,
-                                  .value = keymap_index},
+                                  .value = KM_TAP_HOLD_KEY},
         actual_report);
 
     // assert: tap fires immediately, next event in 50ms
@@ -90,15 +54,12 @@ void test_event_driven_tap_hold_key_tap_release_reported(void) {
   // assemble: init keymap
   keymap_init();
 
-  // Tap the C & TH.LCtrl
-  uint16_t keymap_index = 0;
-
   // act: press tap-hold key at t=0
   {
     uint32_t actual_next_ev_ms = keymap_register_input_after_ms(
         0,
         (struct KeymapInputEvent){.event_type = KeymapEventPress,
-                                  .value = keymap_index},
+                                  .value = KM_TAP_HOLD_KEY},
         actual_report);
 
     // assert: hold timeout scheduled, polling not required yet
@@ -113,7 +74,7 @@ void test_event_driven_tap_hold_key_tap_release_reported(void) {
     uint32_t actual_next_ev_ms = keymap_register_input_after_ms(
         150,
         (struct KeymapInputEvent){.event_type = KeymapEventRelease,
-                                  .value = keymap_index},
+                                  .value = KM_TAP_HOLD_KEY},
         actual_report);
 
     // assert: tap reported, polling required until tap clears
@@ -143,3 +104,7 @@ void test_event_driven_tap_hold_key_tap_release_reported(void) {
     TEST_ASSERT_FALSE(keymap_requires_polling());
   }
 }
+
+#else
+#error "requires SUITE_TAP_HOLD"
+#endif
